Move Party + crypto into dedicated module
