+++ hydra-node/src/Hydra/Network/Ouroboros.hs
@@ -8,9 +8,7 @@ module Hydra.Network.Ouroboros (
 import Hydra.Prelude
 
 import Cardano.Tracing.OrphanInstances.Network ()
-import Codec.CBOR.Term (
-  Term,
- )
+import Codec.CBOR.Term (Term,)
 import qualified Codec.CBOR.Term as CBOR
 import Control.Concurrent.STM (
   TChan,
