+++ hydra-node/src/Hydra/ServerOutput.hs
@@ -3,8 +3,9 @@
 
 module Hydra.ServerOutput where
 
-import Hydra.Ledger (Party, Tx, Utxo, ValidationError)
+import Hydra.Ledger (Tx, Utxo, ValidationError)
 import Hydra.Network (Host)
+import Hydra.Party (Party)
 import Hydra.Prelude
 import Hydra.Snapshot (Snapshot)
